Fix Gradle up-to-date checks on buildCCDXlsx (#1310)
By declaring task inputs and outputs, so that Gradle will not run the
task if the generateCCDConfig task is up to date.

More work remains todo in the config generator to generate deterministic
(ie. sorted) json, so that we only rebuild the xls if there is an actual
config change.