Add new util class to get all notify email templates that match a given search criteria
With this util class we can now search a templates values for a dynamic content variable
or we can search the body of the template for static text.

Useful for when changes are needed to notify email templates we can now find which templates
from the list will need changing.