Use AlternativeServiceOutcome to retrieve the alternativeServiceType
When Legal advisor makes a decision to deny the case, the alternativeService
is set to null and archived inside CaseData.AlternativeServiceOutcomes.

Update logic in aboutToStart and aboutToSubmit to use the most recent outcome
from the list, instead of the now null alterative service.

Update unit tests to cover new code.