NFDIV-2215 - Solicitor Archive Case event (#2580)
* NFDIV-2215 - Solicitor Archive Case event

When a case is in a pre-submitted state, allow a solicitor
to archive the case. Also allow solicitor to progress an
archived case via the "Amend application" & "sign & submit" events.

* Fix style issues

* add solicitor access for state

* Add all pre submission states

* NFDIV-2215 - Removal all access for caseworker

As per new requirements, remove all acess from caseworker for new
archive case event and state.

This may change in future and caseworker access may be required.

* NFDIV-2215 - Remove eventNotes for new solicitor event.

The ccd-case-ui-toolkit doesn't allow solicitors to add or
view eventNotes. This issue was confirmed with a BA, and the decision
was made to remove the eventNotes from the new solicitor event.