suppress snakeyaml cves for now no update (#2782)
* suppress snakeyaml cves for now no update

* Fix failing playwright tests

* Fix failing playwright login

* increase timeout incase this is cause of failures

* revert changes and increase timeout of playwright tests

* comment out code for debugging logs

* delete playwrght code temp for debugging

* re-add code after debugging

* Disable CFTLib playwright tests

For now due to excessive workload, and not able to fix playwright tests
we are removing them. These are covered in e2e tests, function and integration
so nothing will be lost. Eventually these will be re-enabled and more will be implemented

* Re-enable CFTLib tests

Someone on the rse dev tools fixed idam simulator, meaning it wasn't an
issue with our tests or config.