NFD - Move event to all states
We're seeing in production an issue where individual cases are failing
to be unlinked from a bulk case and the error is returning from CCD
as 422 Case status did not qualify for the event. Moving the event to be
available from all states to see if we can quickly fix this issue.