NFD - Add restrictions to bulk case flow (WIP)
Create a new state specifically for cases which are in bulk lists.

Update the caseworker-drop-case-from-bulk-list event to move a case
back to AwaitingPronouncement if a caseworker drops it from a bulk list.
This is to allow the cron which picks cases up and adds them to bulk lists,
to find the case if no other actions are taken against it.