Rework the revert to generate and send at different times
Move the generation of the document into aboutToSubmit but
move the sending of the document into submitted. We do this because
users can upload "document attachments" via XUI who don't use CDAM yet.

This meant that we were getting a 403 when sending in aboutToSubmit as
the documents hadn't been added to the case yet.

Moving to submitted to try and fix this issue.